Move get_serial_by_id and human_readable_device_name to usb
- This code will be reused for zwave_js and other integrations
  that use USB devices.
